New org

+++ hnsw/src/helpers/args.rs
@@ -18,16 +18,16 @@ pub fn parse_args() -> (usize, usize) {
     (dim, lim)
 }
 
-pub fn parse_args_eval() -> Result<(u32, usize, u8), &'static str> {
+pub fn parse_args_eval() -> Result<(usize, usize, usize), &'static str> {
     let args: Vec<String> = env::args().collect();
 
     if args.len() != 4 {
         return Err("Expected exactly 3 positional arguments.");
     }
 
-    let dim = args[1].parse::<u32>().expect("Could not parse dimention");
+    let dim = args[1].parse::<usize>().expect("Could not parse dimention");
     let lim = args[2].parse::<usize>().expect("Could not parse limit");
-    let m = args[3].parse::<u8>().expect("Could not parse M");
+    let m = args[3].parse::<usize>().expect("Could not parse M");
     Ok((dim, lim, m))
 }
